cc: add fast (low overhead) duration timers for linux

+++ benchmarks/bench_storage.h
@@ -17,6 +17,6 @@ rstatus_i bench_storage_deinit(void);
 rstatus_i bench_storage_put(struct benchmark_entry *e);
 rstatus_i bench_storage_get(struct benchmark_entry *e);
 rstatus_i bench_storage_rem(struct benchmark_entry *e);
-size_t bench_storage_config_nopts(void);
+unsigned bench_storage_config_nopts(void);
 void bench_storage_config_init(void *opts);
